requirements.txt
streamlit
langchain-core
langchain-google-genai
langfuse
pydantic
python-dotenv